// frontend/src/client/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

/**
 * Channel
 */
export type Channel = {
    /**
     * Description
     */
    description: string;
    /**
     * Item
     */
    item: Array<Item>;
    /**
     * Link
     */
    link: string;
    /**
     * Title
     */
    title: string;
};

/**
 * HTTPValidationError
 */
export type HttpValidationError = {
    /**
     * Detail
     */
    detail?: Array<ValidationError>;
};

/**
 * Item
 */
export type Item = {
    /**
     * Description
     */
    description: string | null;
    /**
     * Link
     */
    link: string | null;
    /**
     * Pubdate
     */
    pubDate: string;
    /**
     * Title
     */
    title: string | null;
};

/**
 * Rss
 */
export type Rss = {
    channel: Channel;
    /**
     * Version
     */
    version?: string | null;
};

/**
 * ValidationError
 */
export type ValidationError = {
    /**
     * Location
     */
    loc: Array<string | number>;
    /**
     * Message
     */
    msg: string;
    /**
     * Error Type
     */
    type: string;
};

export type RssFeedsApiV1RssGetData = {
    body?: never;
    path?: never;
    query?: {
        /**
         * Until
         */
        until?: string | null;
        /**
         * Tag
         */
        tag?: string | null;
    };
    url: '/api/v1/rss/';
};

export type RssFeedsApiV1RssGetErrors = {
    /**
     * Validation Error
     */
    422: HttpValidationError;
};

export type RssFeedsApiV1RssGetError = RssFeedsApiV1RssGetErrors[keyof RssFeedsApiV1RssGetErrors];

export type RssFeedsApiV1RssGetResponses = {
    /**
     * Response Rss Feeds Api V1 Rss  Get
     * Successful Response
     */
    200: Array<Rss>;
};

export type RssFeedsApiV1RssGetResponse = RssFeedsApiV1RssGetResponses[keyof RssFeedsApiV1RssGetResponses];

export type RssPreviewApiV1RssPreviewGetData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/api/v1/rss/preview';
};

export type RssPreviewApiV1RssPreviewGetResponses = {
    /**
     * Successful Response
     */
    200: unknown;
};

export type RssTagsApiV1RssTagsGetData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/api/v1/rss/tags';
};

export type RssTagsApiV1RssTagsGetResponses = {
    /**
     * Response Rss Tags Api V1 Rss Tags Get
     * Successful Response
     */
    200: Array<string>;
};

export type RssTagsApiV1RssTagsGetResponse = RssTagsApiV1RssTagsGetResponses[keyof RssTagsApiV1RssTagsGetResponses];

export type LivenessLivenessGetData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/liveness';
};

export type LivenessLivenessGetResponses = {
    /**
     * Successful Response
     */
    200: unknown;
};

export type ClientOptions = {
    baseUrl: `${string}://openapi.yaml` | (string & {});
};